// src/app/modules/transportation/routing/model/transport-schedule-view.ts
import { TransportScheduleRoutesView } from './transport-schedule-routes-view';

/**
 * Transport
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * TransportScheduleView
 */
export interface TransportScheduleView { 
    /**
     * Gets Or Sets
     */
    name?: string;
    /**
     * Gets Or Sets
     */
    inboundOrOutbound?: string;
    /**
     * Gets Or Sets
     */
    startTime?: string;
    /**
     * Gets Or Sets
     */
    endTime?: string;
    /**
     * Gets or Sets
     */
    transportScheduleRoutesView?: Array<TransportScheduleRoutesView>;
}